+++ R/DSC_Hierarchical.R
@@ -34,6 +34,9 @@
 #'
 #' @family DSC_Macro
 #'
+#' @param formula `NULL` to use all features in the stream or a model [formula] of the form `~ X1 + X2`
+#'   to specify the features used for clustering. Only `.`, `+` and `-` are currently
+#'   supported in the formula.
 #' @param k The number of desired clusters.
 #' @param h Height where to cut the dendrogram.
 #' @param method the agglomeration method to be used. This should be (an
@@ -75,52 +78,74 @@
 #' update(hc, stream, 500)
 #' plot(hc, stream)
 #' @export
-DSC_Hierarchical <- function(k=NULL, h=NULL, method = "complete",
-  min_weight=NULL, description=NULL) {
-
+DSC_Hierarchical <- function(formula = NULL,
+  k = NULL,
+  h = NULL,
+  method = "complete",
+  min_weight = NULL,
+  description = NULL) {
   hierarchical <- hierarchical$new(
-    k=k, h=h, method=method, min_weight=min_weight)
+    k = k,
+    h = h,
+    method = method,
+    min_weight = min_weight
+  )
 
-  if(is.null(description)) description <- paste("Hierarchical (", method, ")",
-    sep='')
+  if (is.null(description))
+    description <- paste("Hierarchical (", method, ")",
+      sep = '')
 
-  l <- list(description = description, RObj = hierarchical)
+  l <- list(description = description,
+    formula = formula,
+    RObj = hierarchical)
 
-  class(l) <- c("DSC_Hierarchical","DSC_Macro","DSC_R","DSC")
+  class(l) <- c("DSC_Hierarchical", "DSC_Macro", "DSC_R", "DSC")
   l
 }
 
 
 ### calculate centroids
-.centroids <- function(centers, weights, assignment){
-
+.centroids <- function(centers, weights, assignment) {
   macroID <- unique(assignment)
   macroID <- macroID[!is.na(macroID)]
   assignment[is.na(assignment)] <- -1 ### prevent NAs in matching
 
-  cs <- t(sapply(macroID, FUN=
+  cs <- t(sapply(
+    macroID,
+    FUN =
       function(i) {
-        take <- assignment==i
-        colSums(centers[take, ,drop=FALSE] *
-            matrix(weights[take], nrow=sum(take), ncol=ncol(centers))) /
+        take <- assignment == i
+        colSums(centers[take, , drop = FALSE] *
+            matrix(
+              weights[take],
+              nrow = sum(take),
+              ncol = ncol(centers)
+            )) /
           sum(weights[take])
-      }))
+      }
+  ))
 
   ### handle 1-d case
-  if(ncol(centers) == 1) cs <- t(cs)
+  if (ncol(centers) == 1)
+    cs <- t(cs)
   rownames(cs) <- NULL
   colnames(cs) <- colnames(centers)
 
   cs <- data.frame(cs)
 
-  ws <- sapply(macroID, FUN =
-      function(i) sum(weights[assignment==i], na.rm=TRUE))
+  ws <- sapply(
+    macroID,
+    FUN =
+      function(i)
+        sum(weights[assignment == i], na.rm = TRUE)
+  )
 
-  list(centers=cs, weights=ws)
+  list(centers = cs, weights = ws)
 }
 
 
-hierarchical <- setRefClass("hierarchical",
+hierarchical <- setRefClass(
+  "hierarchical",
   fields = list(
     data	= "data.frame",
     dataWeights = "numeric",
@@ -132,22 +157,26 @@ hierarchical <- setRefClass("hierarchical",
     details = "ANY",
     centers	= "data.frame",
     weights = "numeric",
-    min_weight = "numeric"
+    min_weight = "numeric",
+    colnames = "ANY"
   ),
 
   methods = list(
-    initialize = function(
-      k=NULL,
-      h=NULL,
+    initialize = function(k = NULL,
+      h = NULL,
       method	= "complete",
-      min_weight = NULL
-    ) {
+      min_weight = NULL) {
+      if (is.null(k) &&
+          is.null(h))
+        stop("Either h or k needs to be specified.")
+      if (!is.null(k) &&
+          !is.null(h))
+        stop("Only h or k  can be specified.")
 
-      if(is.null(k) && is.null(h)) stop("Either h or k needs to be specified.")
-      if(!is.null(k) && !is.null(h)) stop("Only h or k  can be specified.")
-
-      if(is.null(min_weight)) min_weight <<- 0
-      else min_weight <<- as.numeric(min_weight)
+      if (is.null(min_weight))
+        min_weight <<- 0
+      else
+        min_weight <<- as.numeric(min_weight)
 
       data	<<- data.frame()
       dataWeights	<<- numeric()
@@ -157,6 +186,8 @@ hierarchical <- setRefClass("hierarchical",
       k	<<- k
       h <<- h
 
+      colnames <<- NULL
+
       .self
     }
 
@@ -164,28 +195,28 @@ hierarchical <- setRefClass("hierarchical",
 )
 
 hierarchical$methods(
-  cluster = function(x,  weight = rep(1,nrow(x)), ...) {
+  cluster = function(x,  weight = rep(1, nrow(x)), ...) {
     #if(nrow(x)==1)
     #  warning("DSC_Hierarchical does not support iterative updating! Old data is overwritten.")
 
 
     ### filter weak clusters
-    if(min_weight>0) {
-      x <- x[weight>min_weight,]
-      weight <- weight[weight>min_weight]
+    if (min_weight > 0) {
+      x <- x[weight > min_weight, ]
+      weight <- weight[weight > min_weight]
     }
 
     data <<- x
     dataWeights <<- weight
 
-    if((!is.null(k) && nrow(data) <=k) || nrow(data)<2) {
+    if ((!is.null(k) && nrow(data) <= k) || nrow(data) < 2) {
       centers <<- x
       weights <<- weight
-    }else{
-      hierarchical <- hclust(d=dist(x), method = method)
+    } else{
+      hierarchical <- hclust(d = dist(x), method = method)
       details <<- hierarchical
 
-      if(is.null(k) || k < length(unlist(hierarchical['height'])))
+      if (is.null(k) || k < length(unlist(hierarchical['height'])))
         assignment <<- cutree(hierarchical, k = k, h = h)
       else
         assignment <<- 1
@@ -199,16 +230,29 @@ hierarchical$methods(
     invisible(data.frame(.class = assignment))
   },
 
-  get_microclusters = function(...) { .nodots(...); data },
-  get_microweights = function(...) { .nodots(...); dataWeights },
+  get_microclusters = function(...) {
+    .nodots(...)
+    data
+  },
+  get_microweights = function(...) {
+    .nodots(...)
+    dataWeights
+  },
 
-  get_macroclusters = function(...) { .nodots(...); centers },
-  get_macroweights = function(...) { .nodots(...); weights },
+  get_macroclusters = function(...) {
+    .nodots(...)
+    centers
+  },
+  get_macroweights = function(...) {
+    .nodots(...)
+    weights
+  },
 
-  microToMacro = function(micro=NULL, ...){
-    .nodots(...);
-    if(is.null(micro)) micro <- 1:nrow(data)
-    structure(assignment[micro], names=micro)
+  microToMacro = function(micro = NULL, ...) {
+    .nodots(...)
+
+    if (is.null(micro))
+      micro <- 1:nrow(data)
+    structure(assignment[micro], names = micro)
   }
 )
-